pyproject.toml
[tool.poetry]
name = "redvirt"
version = "0.1.0"
description = "Test task for employment in Red Soft"
authors = ["Arkadii Paktusov <paktusov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
libvirt-python = "^9.5.0"
jsonrpcserver = "^5.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
